Restructured code for better readability
