+++ src/config.js
@@ -1,0 +1,5 @@
+import * as dotenv from 'dotenv'
+
+dotenv.config()
+
+export const { BOT_GIPHY_API_KEY, DESERTORES_CHAT_ID, TELEGRAM_TOKEN } = process.env
